+++ addOns/authhelper/src/main/java/org/zaproxy/addon/authhelper/HeaderBasedSessionManagementMethodType.java
@@ -222,7 +222,7 @@ public class HeaderBasedSessionManagementMethodType extends SessionManagementMet
             } else {
                 LOGGER.debug(
                         "processMessageToMatchSession unexpected session type: {}",
-                        session.getClass().getCanonicalName());
+                        session != null ? session.getClass().getCanonicalName() : "null");
             }
         }