+++ addOns/authhelper/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Fixed
+- Potential timing issue trying to use browser based auth to authenticate before the session management method has been identified.
 
 ## [0.13.0] - 2024-05-07
 ### Changed